Move Main and MainApp to commons
