+++ src/main/java/spleetwaise/address/ui/UiManager.java
@@ -7,7 +7,7 @@ import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
-import spleetwaise.address.MainApp;
+import spleetwaise.commons.MainApp;
 import spleetwaise.commons.core.LogsCenter;
 import spleetwaise.commons.logic.Logic;
 import spleetwaise.commons.util.StringUtil;
