+++ src/test/java/spleetwaise/address/ui/UiPartTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import javafx.fxml.FXML;
-import spleetwaise.address.MainApp;
 import spleetwaise.address.testutil.Assert;
+import spleetwaise.commons.MainApp;
 
 public class UiPartTest {
